requirements.txt
python-dotenv
pytest
pytest-mock
pytest-cov
pre-commit
google-cloud-logging
google-cloud-firestore